Class renamed as per the changes in actual controller classes.
